requirements.txt
asynckivy==0.5.4
beautifulsoup4==4.11.1
certifi==2022.9.24
dill==0.3.6
expiringdict==1.2.2
icalendar==4.1.0
jnius==1.1.0
kivymd==1.0.2
matplotlib==3.6.1
networkx==2.8.8
plyer==2.0.0
pyjnius==1.4.2
requests==2.28.1
